pyproject.toml
[tool.poetry]
name = "slidev2pptx"
version = "0.1.0"
description = ""
authors = ["Yu Zhang <yuzhang94@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-pptx = "^0.6.23"
pillow = "^10.3.0"
click = "^8.1.7"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
slidev2pptx = "slidev2pptx.cli:cli"
